[Refactor] (test) Move the testing *TestZookeeperPath* to unit test.
